Crate bodyparser

Body Parser Plugin for Iron

This plugin parses JSON out of an incoming Request.

Structs

BodyError

Json
Json is a plugin to parse a request body into JSON. Uses Raw plugin to parse the body with limit.

LimitReader
Original impl

MaxBodyLength
Use this key to modify the default body limit.

Raw
Raw is a plugin to read a request body into UTF-8 String. Doesn't read multipart/form-data.

Struct
Struct is a plugin to parse a request body into a struct. Uses Raw plugin to parse the body with limit.

Enums

BodyErrorCause