Struct bodyparser::LimitReader

pub struct LimitReader<R> {
    // some fields omitted
}

Original impl

The LimitReader from the std just stops to read when reaches a limit, but we don't want to return partially readed body to the client code because it is useless. This modified LimitReader returns IoError with IoErrorKind::InvalidInput when it reaches the limit.

Methods

impl<R: Read> LimitReader<R>

fn new(r: R, limit: usize) -> LimitReader<R>

fn into_inner(self) -> R

fn limit(&self) -> usize